session_02/requirements.txt
google-cloud-pubsub==2.12.0
Faker==13.3.4